Fix more signed conversion issues and boostmp conversions

+++ symengine/mp_class.h
@@ -2,6 +2,7 @@
 #define SYMENGINE_INTEGER_CLASS_H
 
 #include <symengine/symengine_config.h>
+#include <symengine/symengine_casts.h>
 #if SYMENGINE_INTEGER_CLASS != SYMENGINE_BOOSTMP
 #include <symengine/mp_wrapper.h>
 #endif
@@ -740,7 +741,7 @@ inline void mp_and(integer_class &res, const integer_class &a,
 inline void mp_pow_ui(integer_class &res, const integer_class &i,
                       unsigned long n)
 {
-    res = boost::multiprecision::pow(i, n);
+    res = boost::multiprecision::pow(i, numeric_cast<unsigned>(n));
 }
 
 inline void mp_gcd(integer_class &res, const integer_class &a,
